+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the EasyAPI plugin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+-  export swagger3 `@Parameter#example` as the query/form/path/header parameter example
+
+---
+
 ## [3.2.0] - 2026-07-05
 
 ### Added